% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwfec.R
\name{convcoder}
\alias{convcoder}
\title{Convolutional Coder
This convcoder documentation}
\usage{
convcoder(x, G)
}
\arguments{
\item{x}{bits received vector of bits (0's and 1's)}

\item{G}{matrix of n code words n rows, K length code words}
}
\value{
ldsfjsldaf
}
\description{
Convolutional Coder
This convcoder documentation
}
\examples{
# Example coder encodes the sequence 1101. The figure illustrates 
# a rate 1/2 (k=1, n=2), constraint length K = 3, convolutional
# code. Convolutional code is described by the codes g1 = (1,1,0) 
# and g2 = (1,1,1). Coded output is 1,1,0,1,1,1,1,0,1,0,0,1
#
#                        (1,1,1) n = 6 5 4 3 2 1
#                  -------> + -->    1 1 0 0 0 1
#                 |         ^
#                 |     ____|
# n = 4 3 2 1     |    |               n =   12 11 10 9 8 7 6 5 4 3 2 1
#   = 1 0 1 1  -> i -> x -> x2         y  =  1  1  0  1 0 0 1 0 1 0 1 1
#                 |         |
#                 |         v         n = 6 5 4 3 2 1 
#                  -------> + -------->   1 0 0 1 1 1
#                        (1,0,1)
#     g1 = 1, 1, 0
#     g2 = 1, 0, 1           
# Example 
  g1=c(1,1,1)
  g2=c(1,0,1)
  x=c(1,1,0,1)  # t= 0 1 2 3 ...
  G=rbind(g1,g2)
  y=convcoder(y,G)
}

